requirements.txt
asgiref==3.7.2
Django==4.2.6
django-phonenumber-field==7.2.0
phonenumbers==8.13.22
sqlparse==0.4.4
tzdata==2023.3
